Make PrestoFileSystemCache accessible
Making PrestoFileSystemCache non final, and adding a method to provide
FileSystemWrapper object.
