+++ src/main/java/org/apache/hadoop/fs/OutputStreamWrapper.java
@@ -1,0 +1,29 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.hadoop.fs;
+
+import java.io.IOException;
+
+public class OutputStreamWrapper
+        extends FSDataOutputStream
+{
+    private final FileSystem fileSystem;
+
+    public OutputStreamWrapper(FSDataOutputStream delegate, FileSystem fileSystem)
+            throws IOException
+    {
+        super(delegate, null, delegate.getPos());
+        this.fileSystem = fileSystem;
+    }
+}
